Механика передачи очков страха и улучшена пуля лечения
При лечении, теперь очки страха передаются от того кого лечат, тому кто лечит.
При смерти моба, теперь очки страха моба передаются убившему его герою.
Замедлена пуля лечения, она стала прикольней.

+++ Assets/Scripts/UI/Scr_HealthBar.cs
@@ -37,7 +37,7 @@ public class Scr_HealthBar : MonoBehaviour
         if (_hero.Health.health != health)
         {
             health = _hero.Health.health;
-            _barNum.text = _hero.Health.health.ToString();
+            _barNum.text = ((int)_hero.Health.health).ToString();
             _bar.fillAmount = health / maxHealth;
         }
     }
